refactor(sps): move BackfillServiceImpl alias from domain to wiring module — keeps domain layer adapter-agnostic
